Improve element visibility checking
In order to check that the element is actually visible we now use
the isDisplayed() function.
I also noticed that the findElement function returns before it
actuyally found the element which means the test could pass even
if the element is not found. To fix this, 'await' and 'assert'
are used.